+++ tokio-net/tests/udp.rs
@@ -1,7 +1,6 @@
 #![warn(rust_2018_idioms)]
 
-use tokio_codec::{Decoder, Encoder};
-use tokio_net::udp::{UdpFramed, UdpSocket};
+use tokio_net::udp::UdpSocket;
 
 use bytes::{BufMut, BytesMut};
 use futures_util::{future::FutureExt, sink::SinkExt, stream::StreamExt, try_future::try_join};
@@ -73,74 +72,5 @@ async fn reunite_error() -> std::io::Result<()> {
     let (s, _) = socket.split();
     let (_, r1) = socket1.split();
     assert!(s.reunite(r1).is_err());
-    Ok(())
-}
-
-pub struct ByteCodec;
-
-impl Decoder for ByteCodec {
-    type Item = Vec<u8>;
-    type Error = io::Error;
-
-    fn decode(&mut self, buf: &mut BytesMut) -> Result<Option<Vec<u8>>, io::Error> {
-        let len = buf.len();
-        Ok(Some(buf.split_to(len).to_vec()))
-    }
-}
-
-impl Encoder for ByteCodec {
-    type Item = Vec<u8>;
-    type Error = io::Error;
-
-    fn encode(&mut self, data: Vec<u8>, buf: &mut BytesMut) -> Result<(), io::Error> {
-        buf.reserve(data.len());
-        buf.put(data);
-        Ok(())
-    }
-}
-
-#[tokio::test]
-async fn send_framed() -> std::io::Result<()> {
-    let mut a_soc = UdpSocket::bind("127.0.0.1:0").await?;
-    let mut b_soc = UdpSocket::bind("127.0.0.1:0").await?;
-
-    let a_addr = a_soc.local_addr()?;
-    let b_addr = b_soc.local_addr()?;
-
-    // test sending & receiving bytes
-    {
-        let mut a = UdpFramed::new(a_soc, ByteCodec);
-        let mut b = UdpFramed::new(b_soc, ByteCodec);
-
-        let msg = b"4567".to_vec();
-
-        let send = a.send((msg.clone(), b_addr));
-        let recv = b.next().map(|e| e.unwrap());
-        let (_, received) = try_join(send, recv).await.unwrap();
-
-        let (data, addr) = received;
-        assert_eq!(msg, data);
-        assert_eq!(a_addr, addr);
-
-        a_soc = a.into_inner();
-        b_soc = b.into_inner();
-    }
-
-    // test sending & receiving an empty message
-    {
-        let mut a = UdpFramed::new(a_soc, ByteCodec);
-        let mut b = UdpFramed::new(b_soc, ByteCodec);
-
-        let msg = b"".to_vec();
-
-        let send = a.send((msg.clone(), b_addr));
-        let recv = b.next().map(|e| e.unwrap());
-        let (_, received) = try_join(send, recv).await.unwrap();
-
-        let (data, addr) = received;
-        assert_eq!(msg, data);
-        assert_eq!(a_addr, addr);
-    }
-
     Ok(())
 }
